python: semantic support (#539)
* Remove meta filter for derived sources objects (#536)

* removed html parser

* renamed SemanticGenerator to JavaSemanticGenerator

* added python semantic support

* commented html parser

* commented html parser

* added html parser

* added n-> -1 semantic

* python sink semantic

* removed println

* updated semantic

* python - apply sink semantic only on leakage till the issue reported is resolved

---------
