+++ src/main/scala/ai/privado/dataflow/Dataflow.scala
@@ -24,9 +24,11 @@
 package ai.privado.dataflow
 
 import ai.privado.cache.{AppCache, AuditCache, DataFlowCache, RuleCache}
-import ai.privado.entrypoint.{PrivadoInput, TimeMetric}
+import ai.privado.entrypoint.{PrivadoInput, ScanProcessor, TimeMetric}
+
 import ai.privado.exporter.ExporterUtility
-import ai.privado.languageEngine.java.semantic.SemanticGenerator
+import ai.privado.languageEngine.java.semantic.JavaSemanticGenerator
+import ai.privado.languageEngine.python.semantic.PythonSemanticGenerator
 import ai.privado.model.{CatLevelOne, Constants, InternalTag, Language}
 import io.joern.dataflowengineoss.language._
 import io.joern.dataflowengineoss.queryengine.{EngineConfig, EngineContext}
@@ -36,6 +38,7 @@ import io.shiftleft.semanticcpg.language._
 import org.slf4j.LoggerFactory
 import overflowdb.traversal.Traversal
 import ai.privado.model.exporter.DataFlowPathIntermediateModel
+import io.joern.dataflowengineoss.semanticsloader.Semantics
 
 import java.util.Calendar
 import scala.collection.mutable.ListBuffer
@@ -46,6 +49,7 @@ class Dataflow(cpg: Cpg) {
   private val logger = LoggerFactory.getLogger(getClass)
 
   /** Compute the flow of data from tagged Sources to Sinks
+    *
     * @return
     *   \- Map of PathId -> Path corresponding to source to sink path
     */
@@ -57,10 +61,7 @@ class Dataflow(cpg: Cpg) {
 
     logger.info("Generating dataflow")
     implicit val engineContext: EngineContext =
-      EngineContext(
-        semantics = SemanticGenerator.getSemantics(cpg, privadoScanConfig, ruleCache),
-        config = EngineConfig(4)
-      )
+      EngineContext(semantics = getSemantics(cpg, privadoScanConfig, ruleCache), config = EngineConfig(4))
     val sources = Dataflow.getSources(cpg)
     val sinks   = Dataflow.getSinks(cpg)
 
@@ -183,6 +184,15 @@ class Dataflow(cpg: Cpg) {
       .toSet
       .map((pathId: String) => (pathId, DataFlowCache.dataflowsMapByType.get(pathId)))
       .toMap
+  }
+
+  def getSemantics(cpg: Cpg, privadoScanConfig: PrivadoInput, ruleCache: RuleCache): Semantics = {
+    val lang = AppCache.repoLanguage
+    lang match {
+      case Language.JAVA   => JavaSemanticGenerator.getSemantics(cpg, privadoScanConfig, ruleCache)
+      case Language.PYTHON => PythonSemanticGenerator.getSemantics(cpg, ruleCache)
+      case _               => JavaSemanticGenerator.getDefaultSemantics
+    }
   }
 
 }
